Test that basic mail fields get set
This unconvered an inconsistency in `from`, which was being set to a
`Mail::Field` rather than a `String`. `Mail::Field#to_s` happens to be
what we want, so it was working through to the SendGrid API, but
speccing was difficult.

The From header can technically contain more than one address, but the
SendGrid API docs don't say that's supported, so we're assuming there's
just one.