Add specs for email body and attachments
Test that all our text and HTML parts are correct, and that attachments
are attached. The body assignment could be simplified, and specs needed
the file to be flushed so we can compare its contents properly - this is
probably the case when sending, too. Lastly, we explicitly close the
Tempfile as encouraged in the documentation.

The Tempfile filename prefix doesn't need randomness in it, as
Tempfile's job is to pick a nonexistent filename for use.