Parse From display name into from_name
Now this gem works with emails of the form `Taco Cat <taco@cat.limo>`,
which previously would fail because the SendGrid API requires `from` to
be a simple email address.

The `mail` gem changed between v2.5.* and v2.6.* to rename
`mail[:from].tree` to `mail[:from].address_list` (which is a much better
name), but ActionMailer 4 requires ~>2.5.3. Rails 4 is now two years
old, and sendgrid-actionmailer is intended for use with ActionMailer, so
the 2.5.* series that ActionMailer requires seems reasonable. In fact,
it might be reasonable to depend on a particular version of
ActionMailer itself, but that's more than we specifically need so far.