Fix attachment tempfiles being closed and deleted before they're used.
This was leading to IOError: closed stream errors when used with
sendgrid-ruby v1.

This changes the tests so this error would have been thrown by still
performing all of the default SendGrid::Client behavior, and only
mocking the HTTP request (in the same way sendgrid-ruby mocks things).
This allows for sendgrid-ruby's internal attachment handling to still
run, which is where this error cropped up with the closed files we were
sending in.