Handle Message With Empty Multipart Content
Actionmailer is happy to have an empty text or html part to a multipart
message. This makes Sendgrid unhappy. Sendgrid will return an error when
attempting to send a message where a specific part of a multipart
message is empty.

This update allows for the behavior of actionmailer to coexist with the
requirements of the Sendgrid API.

                ##
               _[]_
              [____]
          .----'  '----.
      .===|    .==.    |===.
      \   |   /####\   |   /
      /   |   \####/   |   \
      '===|    `""`    |==='
      .===|    .==.    |===.
      \   |   /::::\   |   /
      /   |   \::::/   |   \
      '===|    `""`    |==='
      .===|    .==.    |===.
      \   |   /&&&&\   |   /
      /   |   \&&&&/   |   \
      '===|    `""`    |==='
   jgs    '--.______.--'