Don't parse Hashes passed in Mail::Fields
This adds support for returning the unparsed value from a Mail::Field if it's a
Hash already, instead of converting it to its String representation, doing some
naive text-transformation in an attempt to convert from Ruby formatted hashes
into JSON format, which can then be parsed into a Hash.

We already had the original Hash, so this change skips the whole middle part of
the process.