Don't change contents of values in Hashes (#67)
* Refactor json_parse

This should make the individual steps more obvious instead of cramming
everything into the same line.

* Don't parse Hashes passed in Mail::Fields

This adds support for returning the unparsed value from a Mail::Field if it's a
Hash already, instead of converting it to its String representation, doing some
naive text-transformation in an attempt to convert from Ruby formatted hashes
into JSON format, which can then be parsed into a Hash.

We already had the original Hash, so this change skips the whole middle part of
the process.

* json_parse is expected to use Strings for keys

This matches the default behaviour of `JSON.parse`

Respect the symbolize argument

The default behaviour of json_parse was to use symbols for the keys in the
returned Hash. We should preserve that behaviour even when we're not parsing the
submitted value.

* Demonstrate issue

* Use raw Mail::Fields where possible

This takes advantage of the fact that we've stored Hash objects in Mail::Field's
and that we now use the unparsed values from those Mail::Fields instead of the
parsed, string representation of the original value.

* Remove support for mail gems before 2.7

Older versions of mail assumes that all values in Mail::Field objects are
Strings. We need to pass richer objects, usually Hashes, that we can forward to
the SendGrid API. For example when including variables for dynamic templates:

    mail(
      :to => "someone@example.com",
      :template_id => "d-42",
      :dynamic_variables => {
        :name => "Some One"
      }
    )

Previously, Mail::Field would convert the Hash

    {:dynamic_variables => {:name => "Some One"}}

to its String representation

    "{:dynamic_variables=>{:name=>\"Some One\"}}"

which we would then attempt to convert into something resembling JSON, so we
could parse it back into a Hash with the original values. Unfortunately this
would fail in cases where the value of the variable contains special characters,
in particular `=>`.

* Preserve compatibility with Rubies before 2.5

* Prefer duck-typing over type-checking

We don't really care that the object is Mail::Field as long as we can get an
unparsed value from it.